Split redis from postgres logic
