+++ cadence/server/routes.go
@@ -15,8 +15,8 @@ func routes() *http.ServeMux {
 	r := http.NewServeMux()
 	r.Handle("/api/radiodata/sse", radiodata_sse)
 	r.Handle("/api/search", Search())
-	r.Handle("/api/request/id", RequestID())
-	r.Handle("/api/request/bestmatch", RequestBestMatch())
+	r.Handle("/api/request/id", rateLimit(RequestID()))
+	r.Handle("/api/request/bestmatch", rateLimit(RequestBestMatch()))
 	r.Handle("/api/nowplaying/metadata", NowPlayingMetadata())
 	r.Handle("/api/nowplaying/albumart", NowPlayingAlbumArt())
 	r.Handle("/api/history", History())
